+++ packages/accordion/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@recordreplay/accordion",
+  "version": "0.1.0",
+  "description": "",
+  "main": "index.tsx",
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/RecordReplay/devtools.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/RecordReplay/devtools/issues"
+  },
+  "homepage": "https://github.com/RecordReplay/devtools#readme"
+}
